+++ app/src/main/java/com/example/savushkin/domain/models/ProductOfRequest.kt
@@ -2,6 +2,6 @@ package com.example.savushkin.domain.models
 
 data class ProductOfRequest (
     var codeProduct : String,
-    var quantity : Double,
+    var quantity : Int,
     var numberRequest : Long
 )
